Publish source and simplify ESM build
