+++ package.json
@@ -14,8 +14,10 @@
   },
   "files": [
     "dist",
+    "src",
     "skills/libretto"
   ],
+  "types": "./dist/index.d.ts",
   "bin": {
     "libretto": "./dist/cli/index.js",
     "libretto-cli": "./dist/cli/index.js"
@@ -24,14 +26,12 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
+      "default": "./dist/index.js"
     }
   },
   "scripts": {
     "postinstall": "playwright install chromium",
-    "build": "pnpm run build:runtime && pnpm run build:cli",
-    "build:runtime": "tsup --config tsup.config.ts",
-    "build:cli": "tsup --config tsup.cli.config.ts",
+    "build": "tsup --config tsup.config.ts",
     "type-check": "tsc --noEmit",
     "test": "pnpm run build && vitest run",
     "eval": "pnpm run build && vitest run --config vitest.evals.config.ts",
